+++ Assets/Dust/Scripts/Runtime/Actions/MoveAction.cs
@@ -6,18 +6,32 @@ namespace DustEngine
     {
         protected Vector3 m_DeltaLocalMove;
 
+        protected bool m_AutoRotateBySelfDirection = false;
+
         //--------------------------------------------------------------------------------------------------------------
         // Dust.Action lifecycle
-        
+
+        protected override void OnActionStart()
+        {
+            base.OnActionStart();
+
+            m_AutoRotateBySelfDirection = false;
+        }
+
         protected override void OnActionUpdate(float deltaTime)
         {
             if (Dust.IsNull(activeTargetTransform))
                 return;
+            
+            Vector3 moveStep = m_DeltaLocalMove * (playbackState - previousState);
+
+            if (m_AutoRotateBySelfDirection)
+                moveStep = DuMath.RotatePoint(moveStep, activeTargetTransform.localEulerAngles);
 
             if (playingPhase == PlayingPhase.Main)
-                activeTargetTransform.localPosition += m_DeltaLocalMove * (playbackState - previousState);
+                activeTargetTransform.localPosition += moveStep;
             else
-                activeTargetTransform.localPosition -= m_DeltaLocalMove * (playbackState - previousState);
+                activeTargetTransform.localPosition -= moveStep;
         }
     }
 }